Abstract query resolvers to separate files
